sentence_transformer_flask_app_bulid/requirements.txt
Flask
Werkzeug
sentence-transformers
PyMuPDF
python-docx
scikit-learn
pandas
numpy